pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Eduardo da Silva", email = "eduardo.silva@gmail.com"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.9.0",
    "isort>=5.12.0",
    "pylint>=2.17.5",
]
